+++ src/core/injections/semiblind/techniques/tempfile_based/tfb_handler.py
@@ -23,7 +23,6 @@ import random
 import base64
 import urllib
 import urllib2
-import readline
   
 from src.utils import menu
 from src.utils import logs
@@ -42,6 +41,23 @@ from src.core.injections.semiblind.techniques.tempfile_based import tfb_enumerat
 from src.core.injections.semiblind.techniques.tempfile_based import tfb_file_access
 
 from src.core.injections.semiblind.techniques.file_based import fb_injector
+
+readline_error = False
+try:
+  import readline
+except ImportError:
+  if settings.IS_WINDOWS:
+    try:
+      import pyreadline as readline
+    except ImportError:
+      readline_error = True
+  else:
+    try:
+      import gnureadline as readline
+    except ImportError:
+      readline_error = True
+  pass
+
 
 """
 The "tempfile-based" injection technique on semiblind OS command injection.
@@ -375,11 +391,19 @@ def tfb_injection_handler(url, delay, filename, tmp_path, http_request_method, u
                   if gotshell in settings.CHOISE_YES:
                     print ""
                     print "Pseudo-Terminal (type '" + Style.BRIGHT + "?" + Style.RESET_ALL + "' for available options)"
+                    if readline_error:
+                      checks.no_readline_module()
                     while True:
                       try:
                         # Tab compliter
-                        readline.set_completer(menu.tab_completer)
-                        readline.parse_and_bind("tab: complete")
+                        if not readline_error:
+                          readline.set_completer(menu.tab_completer)
+                          # MacOSX tab compliter
+                          if 'libedit' in readline.__doc__:
+                            readline.parse_and_bind("bind ^I rl_complete")
+                          # Unix tab compliter
+                          else:
+                            readline.parse_and_bind("tab: complete")
                         cmd = raw_input("""commix(""" + Style.BRIGHT + Fore.RED + """os_shell""" + Style.RESET_ALL + """) > """)
                         cmd = checks.escaped_cmd(cmd)
                         if cmd.lower() in settings.SHELL_OPTIONS: